requirements.txt
# requirements.txt
wheel
flask
pytesseract
pillow

